renamed appname to DailyCommuter and all its mentions
