+++ DailyCommuter/templates/base.html
@@ -3,7 +3,7 @@
 <link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}">
 <nav>
   <!-- change whenever we change the name of the app -->
-  <h1>Appname</h1>
+  <h1>DailyCommuter</h1>
   <!-- Add login page options when implementing users/login -->
 </nav>
 <section class="content">
