+++ pyproject.toml
@@ -1,6 +1,5 @@
 [project]
-# change appname to whatever we rename the app
-name = "appname"
+name = "DailyCommuter"
 version = "0.0.1"
 authors = [
   { name="Conor Zhang", email="chz9577@nyu.edu" },
